Added Refreshere,RecyclerView and CardView to the MainActivity
Implemented a first sketch of a MatchAdapter and modified some layout values. This commit is heavy due to the huge amount of changes in layout visualization. Further work will polish the code of the layouts and will rearrange the CardView to make them look much nicer than the current version.