Huge update
Dota 7.07 patch notes came out yesterday, so i made this huge update as well.
Implemented Firebase in order to get a presence system.
Addeed a new Activity (UserListActivity) that shows the currently connected users.
New layout for the activity AND just noticed that i was missing a menu item on git so i added that aswell. Now the app needs some graphical tweaks but it is done! (At least for my exam)